+++ tests/functional/test_wps_package.py
@@ -112,30 +112,30 @@ class WpsPackageAppTest(unittest.TestCase):
     def test_literal_io_from_package_and_offering(self):
         raise NotImplementedError
 
-    def test_complex_io_with_multiple_formats(self):
-        # FIXME: multiple output (array) not tested here
+    def test_complex_io_with_multiple_formats_and_defaults(self):
         """
         Test validates that different format types are set on different I/O variations simultaneously:
-            - input with 1 format, single value, no default
-            - input with 1 format, array values, no default
-            - input with 1 format, single value, 1 default
-            - input with 1 format, array values, 1 default
-            - input with many formats, single value, no default
-            - input with many formats, array values, no default
-            - input with many formats, single value, 1 default
-            - input with many formats, array values, 1 default
+            - input with 1 format, single value, no default value
+            - input with 1 format, array values, no default value
+            - input with 1 format, single value, 1 default value
+            - input with 1 format, array values, 1 default value
+            - input with many formats, single value, no default value
+            - input with many formats, array values, no default value
+            - input with many formats, single value, 1 default value
+            - input with many formats, array values, 1 default value
+            - output with 1 format, single value
+            - output with 1 format, array values
+            - output with many formats, single value
+            - output with many formats, array values
 
-            - output with 1 format, single value, no default
-            - output with 1 format, single value, 1 default
-            - output with many formats, single value, no default
-            - output with many formats, single value, 1 default
-
-        On addition, the test evaluates that:
+        In addition, the test evaluates that:
             - CWL I/O specified as list preserves the specified ordering
+            - CWL 'default' "value" doesn't interfere with WPS 'default' "format" and vice-versa
             - partial WPS definition of I/O format to indicate 'default' are resolved with additional CWL I/O formats
+            - min/max occurrences are solved accordingly to single/array values and 'default' if not overridden by WPS
 
         NOTE:
-            field 'default' in CWL refers to default "value", in WPS refers to default "format" for complex types
+            field 'default' in CWL refers to default "value", in WPS refers to default "format" for complex inputs
         """
         ns1, type1 = get_cwl_file_format(CONTENT_TYPE_APP_JSON)
         ns2, type2 = get_cwl_file_format(CONTENT_TYPE_TEXT_PLAIN)
@@ -185,7 +185,7 @@ class WpsPackageAppTest(unittest.TestCase):
                         "type": "array",
                         "items": "File",
                     },
-                    "format": [type1, type2, type3],
+                    "format": [type3, type2, type1],
                 },
                 {
                     "id": "single_value_multi_format_default",
@@ -204,14 +204,34 @@ class WpsPackageAppTest(unittest.TestCase):
                 },
             ],
             "outputs": [
-                # FIXME: implement other variations of outputs (see docstring)
                 {
-                    "id": "values",
-                    "type": {
-                        "type": "array",
-                        "items": "float",
-                    }
-                }
+                    "id": "single_value_single_format",
+                    "type": "File",
+                    "format": type1,
+                },
+                # FIXME: multiple output (array) not implemented (https://github.com/crim-ca/weaver/issues/25)
+                # {
+                #    "id": "multi_value_single_format",
+                #    "type": {
+                #        "type": "array",
+                #        "items": "File",
+                #    },
+                #    "format": type1,
+                # },
+                {
+                    "id": "single_value_multi_format",
+                    "type": "File",
+                    "format": [type1, type2, type3]
+                },
+                # FIXME: multiple output (array) not implemented (https://github.com/crim-ca/weaver/issues/25)
+                # {
+                #     "id": "multi_value_multi_format",
+                #     "type": {
+                #         "type": "array",
+                #         "items": "File",
+                #     },
+                #     "format": [type1, type2, type3],
+                # },
             ],
             "$namespaces": namespaces
         }
@@ -228,7 +248,7 @@ class WpsPackageAppTest(unittest.TestCase):
                             "id": "multi_value_multi_format",
                             "formats": [
                                 {
-                                    "mimeType": CONTENT_TYPE_APP_JSON,
+                                    "mimeType": CONTENT_TYPE_TEXT_PLAIN,
                                     "default": True,
                                 }
                             ]
@@ -279,11 +299,11 @@ class WpsPackageAppTest(unittest.TestCase):
         assert desc["process"]["inputs"][4]["minOccurs"] == "1"
         assert desc["process"]["inputs"][4]["maxOccurs"] == "1"
         assert len(desc["process"]["inputs"][4]["formats"]) == 3
-        assert desc["process"]["inputs"][4]["formats"][0]["mimeType"] == CONTENT_TYPE_APP_JSON
-        assert desc["process"]["inputs"][4]["formats"][0]["default"] is True  # no explicit default, uses first
+        assert desc["process"]["inputs"][4]["formats"][0]["mimeType"] == CONTENT_TYPE_APP_NETCDF
+        assert desc["process"]["inputs"][4]["formats"][0]["default"] is False
         assert desc["process"]["inputs"][4]["formats"][1]["mimeType"] == CONTENT_TYPE_TEXT_PLAIN
-        assert desc["process"]["inputs"][4]["formats"][1]["default"] is False
-        assert desc["process"]["inputs"][4]["formats"][2]["mimeType"] == CONTENT_TYPE_APP_NETCDF
+        assert desc["process"]["inputs"][4]["formats"][1]["default"] is True  # no explicit default, uses first
+        assert desc["process"]["inputs"][4]["formats"][2]["mimeType"] == CONTENT_TYPE_APP_JSON
         assert desc["process"]["inputs"][4]["formats"][2]["default"] is False
         assert desc["process"]["inputs"][5]["id"] == "multi_value_multi_format"
         assert desc["process"]["inputs"][5]["minOccurs"] == "1"
@@ -320,13 +340,101 @@ class WpsPackageAppTest(unittest.TestCase):
         # FIXME: implement
 
         # package input validation
-        assert pkg["inputs"]["single_value_single_format"]["format"] == type1
-        assert pkg["inputs"]["single_value_single_format"]["type"] == "File"
-        assert pkg["inputs"]["multi_value_single_format"]["format"] == type1
-        assert pkg["inputs"]["multi_value_single_format"]["type"] == "array"
+        assert pkg["inputs"][0]["id"] == "single_value_single_format"
+        assert pkg["inputs"][0]["type"] == "File"
+        assert pkg["inputs"][0]["format"] == type1
+        assert "default" not in pkg["inputs"][0]
+        assert pkg["inputs"][1]["id"] == "multi_value_single_format"
+        assert pkg["inputs"][1]["type"] == "array"
+        assert pkg["inputs"][1]["items"] == "File"
+        assert pkg["inputs"][1]["id"]["format"] == type2
+        assert "default" not in pkg["inputs"][1]
+        assert pkg["inputs"][2]["id"] == "single_value_single_format_default"
+        assert pkg["inputs"][2]["type"] == "File"
+        assert pkg["inputs"][2]["format"] == type3
+        assert pkg["inputs"][2]["default"] == default_file
+        assert pkg["inputs"][3]["id"] == "multi_value_single_format_default"
+        assert pkg["inputs"][3]["type"] == "array"
+        assert pkg["inputs"][3]["items"] == "File"
+        assert pkg["inputs"][3]["id"]["format"] == type2
+        assert pkg["inputs"][3]["default"] == default_file
+        assert pkg["inputs"][4]["id"] == "multi_value_single_format"
+        assert pkg["inputs"][4]["type"] == "File"
+        assert pkg["inputs"][4]["format"] == [type1, type2, type3]
+        assert "default" not in pkg["inputs"][4]
+        assert pkg["inputs"][5]["id"] == "multi_value_multi_format"
+        assert pkg["inputs"][5]["type"] == "array"
+        assert pkg["inputs"][5]["items"] == "File"
+        assert pkg["inputs"][5]["format"] == [type3, type2, type1]
+        assert "default" not in pkg["inputs"][5]
+        assert pkg["inputs"][6]["id"] == "single_value_multi_format_default"
+        assert pkg["inputs"][6]["type"] == "File"
+        assert pkg["inputs"][6]["format"] == [type1, type2, type3]
+        assert pkg["inputs"][6]["default"] == default_file
+        assert pkg["inputs"][7]["id"] == "single_value_multi_format_default"
+        assert pkg["inputs"][7]["type"] == "array"
+        assert pkg["inputs"][7]["items"] == "File"
+        assert pkg["inputs"][7]["format"] == [type1, type2, type3]
+        assert pkg["inputs"][7]["default"] == default_file
 
         # package output validation
         # FIXME: implement
+
+    # FIXME: implement, test should validate that
+    #   - min=0 if default value, min=1 otherwise when only resolved by CWL
+    #   - max=1 if single value, max="unbounded" if array when only resolved by CWL
+    #   - min AND max each overridden by WPS sets the corresponding value, regardless of single/array CWL resolution
+    @pytest.mark.xfail(reason="not implemented")
+    def test_merging_io_min_max_occurs(self):
+        raise NotImplementedError
+
+    def test_literal_io_from_package(self):
+        cwl = {
+            "cwlVersion": "v1.0",
+            "class": "CommandLineTool",
+            "inputs": {
+                "url": {
+                    "type": "string"
+                }
+            },
+            "outputs": {
+                "values": {
+                    "type": {
+                        "type": "array",
+                        "items": "float",
+                    }
+                }
+            }
+        }
+        body = {
+            "processDescription": {
+                "process": {
+                    "id": self.__name__,
+                    "title": "some title",
+                    "abstract": "this is a test",
+                }
+            },
+            "deploymentProfileName": "http://www.opengis.net/profiles/eoc/wpsApplication",
+            "executionUnit": [{"unit": cwl}],
+        }
+        desc, pkg = self.deploy_process(body)
+        assert desc["process"]["id"] == self.__name__
+        assert desc["process"]["title"] == "some title"
+        assert desc["process"]["abstract"] == "this is a test"
+        assert isinstance(desc["process"]["inputs"], list)
+        assert len(desc["process"]["inputs"]) == 1
+        assert desc["process"]["inputs"][0]["id"] == "url"
+        assert desc["process"]["inputs"][0]["minOccurs"] == "1"
+        assert desc["process"]["inputs"][0]["maxOccurs"] == "1"
+        assert "format" not in desc["process"]["inputs"][0]
+        assert isinstance(desc["process"]["outputs"], list)
+        assert len(desc["process"]["outputs"]) == 1
+        assert desc["process"]["outputs"][0]["id"] == "values"
+        assert "minOccurs" not in desc["process"]["outputs"][0]
+        assert "maxOccurs" not in desc["process"]["outputs"][0]
+        assert "format" not in desc["process"]["outputs"][0]
+        expected_fields = {"id", "title", "abstract", "inputs", "outputs", "executeEndpoint"}
+        assert len(set(desc["process"].keys()) - expected_fields) == 0
 
     # FIXME: implement
     @pytest.mark.xfail(reason="not implemented")
@@ -354,23 +462,29 @@ class WpsPackageAppTest(unittest.TestCase):
         # package I/O validation
         assert "inputs" in pkg
         assert len(pkg["inputs"]) == 2
-        assert pkg["inputs"]["tasmax"]["default"]["mimeType"] == CONTENT_TYPE_APP_NETCDF
-        assert pkg["inputs"]["tasmax"]["default"]["encoding"] == "base64"
-        assert pkg["inputs"]["tasmax"]["default"]["schema"] is None
-        assert pkg["inputs"]["tasmax"]["format"] == EDAM_NETCDF
-        assert pkg["inputs"]["tasmax"]["type"]["type"] == "array"
-        assert pkg["inputs"]["tasmax"]["type"]["items"] == "File"
-        assert pkg["inputs"]["freq"]["default"] == "YS"
-        assert pkg["inputs"]["freq"]["type"]["type"] == "enum"
-        assert pkg["inputs"]["freq"]["type"]["symbols"] == ["YS", "MS", "QS-DEC", "AS-JUL"]
+        assert isinstance(pkg["inputs"], list)
+        assert pkg["inputs"][0]["id"] == "tasmax"
+        assert pkg["inputs"][0]["default"]["mimeType"] == CONTENT_TYPE_APP_NETCDF
+        assert pkg["inputs"][0]["default"]["encoding"] == "base64"
+        assert pkg["inputs"][0]["default"]["schema"] is None
+        assert pkg["inputs"][0]["format"] == EDAM_NETCDF
+        assert pkg["inputs"][0]["type"]["type"] == "array"
+        assert pkg["inputs"][0]["type"]["items"] == "File"
+        assert pkg["inputs"][1]["id"] == "freq"
+        assert pkg["inputs"][1]["default"] == "YS"
+        assert pkg["inputs"][1]["type"]["type"] == "enum"
+        assert pkg["inputs"][1]["type"]["symbols"] == ["YS", "MS", "QS-DEC", "AS-JUL"]
         assert "outputs" in pkg
         assert len(pkg["outputs"]) == 2
-        assert pkg["outputs"]["output_netcdf"]["format"] == EDAM_NETCDF
-        assert pkg["outputs"]["output_netcdf"]["type"] == "File"
-        assert pkg["outputs"]["output_netcdf"]["outputBinding"]["glob"] == "output_netcdf.nc"
-        assert pkg["outputs"]["output_log"]["format"] == EDAM_PLAIN
-        assert pkg["outputs"]["output_log"]["type"] == "File"
-        assert pkg["outputs"]["output_log"]["outputBinding"]["glob"] == "output_log.*"
+        assert isinstance(pkg["outputs"], list)
+        assert pkg["outputs"][0]["id"] == "output_netcdf"
+        assert pkg["outputs"][0]["format"] == EDAM_NETCDF
+        assert pkg["outputs"][0]["type"] == "File"
+        assert pkg["outputs"][0]["outputBinding"]["glob"] == "output_netcdf.nc"
+        assert pkg["outputs"][1]["id"] == "output_log"
+        assert pkg["outputs"][1]["format"] == EDAM_PLAIN
+        assert pkg["outputs"][1]["type"] == "File"
+        assert pkg["outputs"][1]["outputBinding"]["glob"] == "output_log.*"
 
         # process description I/O validation
         assert len(desc["process"]["inputs"]) == 2
@@ -428,22 +542,29 @@ class WpsPackageAppTest(unittest.TestCase):
         # package I/O validation
         assert "inputs" in pkg
         assert len(pkg["inputs"]) == 3
-        assert pkg["inputs"]["region"]["default"] == "DEU"
-        assert pkg["inputs"]["region"]["type"]["type"] == "array"
-        assert pkg["inputs"]["region"]["type"]["items"]["type"] == "enum"
-        assert isinstance(pkg["inputs"]["region"]["type"]["items"]["symbols"], list)
-        assert len(pkg["inputs"]["region"]["type"]["items"]["symbols"]) == 220
-        assert all(isinstance(s, six.string_types) for s in pkg["inputs"]["region"]["type"]["items"]["symbols"])
-        assert pkg["inputs"]["mosaic"]["default"] == "null"
-        assert pkg["inputs"]["mosaic"]["type"] == "boolean"
-        assert pkg["inputs"]["resource"]["default"]["mimeType"] == CONTENT_TYPE_APP_NETCDF
-        assert pkg["inputs"]["resource"]["default"]["encoding"] is None
-        assert pkg["inputs"]["resource"]["default"]["schema"] is None
-        assert pkg["inputs"]["resource"]["format"] == EDAM_NETCDF
-        assert pkg["inputs"]["resource"]["type"]["type"] == "array"
-        assert pkg["inputs"]["resource"]["type"]["items"] == "File"
-        # FIXME: implement (resource input had 3 possible formats)
-        #   figure out how to generate the CWL with multi input formats... is it possible?
+        assert isinstance(pkg["inputs"], list)
+        assert pkg["inputs"][0]["id"] == "region"
+        assert pkg["inputs"][0]["default"] == "DEU"
+        assert pkg["inputs"][0]["type"]["type"] == "array"
+        assert pkg["inputs"][0]["type"]["items"]["type"] == "enum"
+        assert isinstance(pkg["inputs"][0]["type"]["items"]["symbols"], list)
+        assert len(pkg["inputs"][0]["type"]["items"]["symbols"]) == 220
+        assert all(isinstance(s, six.string_types) for s in pkg["inputs"][0]["type"]["items"]["symbols"])
+        assert pkg["inputs"][1]["id"] == "mosaic"
+        assert pkg["inputs"][1]["default"] == "null"
+        assert pkg["inputs"][1]["type"] == "boolean"
+        assert pkg["inputs"][2]["id"] == "resource"
+        assert pkg["inputs"][2]["default"]["mimeType"] == CONTENT_TYPE_APP_NETCDF
+        assert pkg["inputs"][2]["default"]["encoding"] is None
+        assert pkg["inputs"][2]["default"]["schema"] is None
+        assert pkg["inputs"][2]["format"] == EDAM_NETCDF
+        assert pkg["inputs"][2]["type"]["type"] == "array"
+        assert pkg["inputs"][2]["type"]["items"] == "File"
+        assert isinstance(pkg["inputs"][2]["format"], list)
+        assert len(pkg["inputs"][2]["format"]) == 3
+        assert pkg["inputs"][2]["format"][0] == CONTENT_TYPE_APP_NETCDF
+        assert pkg["inputs"][2]["format"][1] == CONTENT_TYPE_APP_TAR
+        assert pkg["inputs"][2]["format"][2] == CONTENT_TYPE_APP_ZIP
 
         # process description I/O validation
         assert len(desc["process"]["inputs"]) == 3