+++ weaver/config.py
@@ -6,7 +6,7 @@ LOGGER = logging.getLogger(__name__)
 WEAVER_CONFIGURATION_DEFAULT = "DEFAULT"
 WEAVER_CONFIGURATION_ADES = "ADES"
 WEAVER_CONFIGURATION_EMS = "EMS"
-weaver_CONFIGURATIONS = frozenset([
+WEAVER_CONFIGURATIONS = frozenset([
     WEAVER_CONFIGURATION_DEFAULT,
     WEAVER_CONFIGURATION_ADES,
     WEAVER_CONFIGURATION_EMS,
@@ -19,7 +19,7 @@ def get_weaver_configuration(settings):
         LOGGER.warn("Setting 'weaver.configuration' not specified, using '{}'".format(WEAVER_CONFIGURATION_DEFAULT))
         weaver_config = WEAVER_CONFIGURATION_DEFAULT
     weaver_config_up = weaver_config.upper()
-    if weaver_config_up not in weaver_CONFIGURATIONS:
+    if weaver_config_up not in WEAVER_CONFIGURATIONS:
         raise ConfigurationError("Unknown setting 'weaver.configuration' specified: '{}'".format(weaver_config))
     return weaver_config_up