+++ weaver/visibility.py
@@ -1,5 +1,5 @@
-VISIBILITY_PUBLIC = 'public'
-VISIBILITY_PRIVATE = 'private'
+VISIBILITY_PUBLIC = "public"
+VISIBILITY_PRIVATE = "private"
 
 visibility_values = frozenset([
     VISIBILITY_PUBLIC,